Enable the app to be build and opened successfully
Updated tauri and plugins, along with specta and add specta_typescript
so we can ensure types are only build and published in development.
Removed some capabilities that seemed to no longer be relevent.
Make no unused locals false in tsconfig.json so that unused functions
in the auto generated  bindings.ts file don't trigger warnings.
Refactor the code to load countries to access the app resources.